# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Astro Interpretador RAG API Strict",
    "description": "Esquema estricto para generaci\u00f3n de clientes",
    "version": "1.0.0"
  },
  "paths": {
    "/interpretar": {
      "post": {
        "summary": "Generar Interpretacion",
        "operationId": "generar_interpretacion_interpretar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretacionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interpretar-eventos": {
      "post": {
        "summary": "Interpretar Eventos Calendario",
        "operationId": "interpretar_eventos_calendario_interpretar_eventos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretacionEventoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretacionEventosResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CartaNatalData": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre",
            "description": "Nombre del consultante"
          },
          "points": {
            "additionalProperties": true,
            "type": "object",
            "title": "Points",
            "description": "Datos de planetas y puntos"
          },
          "houses": {
            "additionalProperties": true,
            "type": "object",
            "title": "Houses",
            "description": "Datos de casas"
          },
          "aspects": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Aspects",
            "description": "Lista de aspectos"
          },
          "cuspides_cruzadas": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuspides Cruzadas"
          },
          "aspectos_cruzados": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspectos Cruzados"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "points",
          "houses",
          "aspects"
        ],
        "title": "CartaNatalData",
        "description": "Datos de carta natal recibidos del microservicio de c\u00e1lculo"
      },
      "EventoCalendario": {
        "properties": {
          "fecha_utc": {
            "type": "string",
            "title": "Fecha Utc"
          },
          "hora_utc": {
            "type": "string",
            "title": "Hora Utc"
          },
          "tipo_evento": {
            "type": "string",
            "title": "Tipo Evento"
          },
          "descripcion": {
            "type": "string",
            "title": "Descripcion"
          },
          "planeta1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planeta1"
          },
          "planeta2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planeta2"
          },
          "posicion1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posicion1"
          },
          "posicion2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posicion2"
          },
          "tipo_aspecto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo Aspecto"
          },
          "orbe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orbe"
          },
          "es_aplicativo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Es Aplicativo"
          },
          "harmony": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harmony"
          },
          "elevacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevacion",
            "default": ""
          },
          "azimut": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Azimut",
            "default": ""
          },
          "signo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signo"
          },
          "grado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grado"
          },
          "posicion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posicion"
          },
          "casa_natal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casa Natal"
          },
          "house_transits": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "House Transits"
          },
          "interpretacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpretacion"
          }
        },
        "type": "object",
        "required": [
          "fecha_utc",
          "hora_utc",
          "tipo_evento",
          "descripcion"
        ],
        "title": "EventoCalendario",
        "description": "Representa un \u00fanico evento de calendario que necesita interpretaci\u00f3n.\nEs un espejo del modelo AstroEventResponse del servicio de calendario."
      },
      "EventoInterpretado": {
        "properties": {
          "descripcion": {
            "type": "string",
            "title": "Descripcion"
          },
          "interpretacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpretacion"
          }
        },
        "type": "object",
        "required": [
          "descripcion"
        ],
        "title": "EventoInterpretado",
        "description": "Representa un evento de calendario con su interpretaci\u00f3n."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterpretacionEventoRequest": {
        "properties": {
          "eventos": {
            "items": {
              "$ref": "#/components/schemas/EventoCalendario"
            },
            "type": "array",
            "title": "Eventos"
          }
        },
        "type": "object",
        "required": [
          "eventos"
        ],
        "title": "InterpretacionEventoRequest",
        "description": "Request para interpretar una lista de eventos de calendario."
      },
      "InterpretacionEventosResponse": {
        "properties": {
          "interpretaciones": {
            "items": {
              "$ref": "#/components/schemas/EventoInterpretado"
            },
            "type": "array",
            "title": "Interpretaciones"
          },
          "tiempo_generacion": {
            "type": "number",
            "title": "Tiempo Generacion"
          }
        },
        "type": "object",
        "required": [
          "interpretaciones",
          "tiempo_generacion"
        ],
        "title": "InterpretacionEventosResponse",
        "description": "Respuesta con una lista de eventos y sus interpretaciones."
      },
      "InterpretacionItem": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "interpretacion": {
            "type": "string",
            "title": "Interpretacion"
          },
          "planeta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planeta"
          },
          "signo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signo"
          },
          "casa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casa"
          },
          "aspecto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspecto"
          },
          "planeta1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planeta1"
          },
          "planeta2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planeta2"
          },
          "grados": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grados"
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "tipo",
          "interpretacion"
        ],
        "title": "InterpretacionItem",
        "description": "Item individual de interpretaci\u00f3n"
      },
      "InterpretacionRequest": {
        "properties": {
          "carta_natal": {
            "$ref": "#/components/schemas/CartaNatalData"
          },
          "genero": {
            "type": "string",
            "title": "Genero",
            "description": "G\u00e9nero: masculino o femenino"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo",
            "description": "Tipo de carta: tropical o draco",
            "default": "tropical"
          }
        },
        "type": "object",
        "required": [
          "carta_natal",
          "genero"
        ],
        "title": "InterpretacionRequest",
        "description": "Request para generar interpretaci\u00f3n"
      },
      "InterpretacionResponse": {
        "properties": {
          "interpretacion_narrativa": {
            "type": "string",
            "title": "Interpretacion Narrativa"
          },
          "interpretaciones_individuales": {
            "items": {
              "$ref": "#/components/schemas/InterpretacionItem"
            },
            "type": "array",
            "title": "Interpretaciones Individuales"
          },
          "tiempo_generacion": {
            "type": "number",
            "title": "Tiempo Generacion"
          }
        },
        "type": "object",
        "required": [
          "interpretacion_narrativa",
          "interpretaciones_individuales",
          "tiempo_generacion"
        ],
        "title": "InterpretacionResponse",
        "description": "Respuesta con interpretaciones completas"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}